refactor: switch to v1beta4 testutil
