+++ go/cli/bme_suite_test.go
@@ -21,7 +21,7 @@ import (
 	clitestutil "pkg.akt.dev/go/cli/testutil"
 	bme "pkg.akt.dev/go/node/bme/v1"
 	"pkg.akt.dev/go/sdkutil"
-	"pkg.akt.dev/go/testutil"
+	testutil "pkg.akt.dev/go/testutil/v1beta4"
 )
 
 type BMECLITestSuite struct {
