+++ go/cli/crisis_tx_test.go
@@ -20,7 +20,7 @@ import (
 	cflags "pkg.akt.dev/go/cli/flags"
 	clitestutil "pkg.akt.dev/go/cli/testutil"
 	"pkg.akt.dev/go/sdkutil"
-	"pkg.akt.dev/go/testutil"
+	testutil "pkg.akt.dev/go/testutil/v1beta4"
 )
 
 func TestNewMsgVerifyInvariantTxCmd(t *testing.T) {
