+++ go/cli/oracle_suite_test.go
@@ -23,7 +23,7 @@ import (
 	clitestutil "pkg.akt.dev/go/cli/testutil"
 	oracle "pkg.akt.dev/go/node/oracle/v1"
 	"pkg.akt.dev/go/sdkutil"
-	"pkg.akt.dev/go/testutil"
+	testutil "pkg.akt.dev/go/testutil/v1beta4"
 )
 
 type OracleCLITestSuite struct {
