+++ go/cli/upgrade_query_test.go
@@ -17,7 +17,7 @@ import (
 
 	"pkg.akt.dev/go/cli"
 	"pkg.akt.dev/go/sdkutil"
-	"pkg.akt.dev/go/testutil"
+	testutil "pkg.akt.dev/go/testutil/v1beta4"
 )
 
 func TestGetCurrentPlanCmd(t *testing.T) {
