+++ go/manifest/v2beta3/manifest_test.go
@@ -10,7 +10,7 @@ import (
 
 	dtypes "pkg.akt.dev/go/node/deployment/v1beta5"
 	resources "pkg.akt.dev/go/node/types/resources/v1beta4"
-	"pkg.akt.dev/go/testutil"
+	testutil "pkg.akt.dev/go/testutil/v1beta4"
 )
 
 const (
