+++ go/node/deployment/v1beta5/msgs_test.go
@@ -9,7 +9,7 @@ import (
 	v1 "pkg.akt.dev/go/node/deployment/v1"
 	types "pkg.akt.dev/go/node/deployment/v1beta5"
 	deposit "pkg.akt.dev/go/node/types/deposit/v1"
-	"pkg.akt.dev/go/testutil"
+	testutil "pkg.akt.dev/go/testutil/v1beta4"
 )
 
 type testMsg struct {
