+++ go/testutil/v1beta3/deployment.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	dtypes "pkg.akt.dev/go/node/deployment/v1"
-	dtypesv1beta4 "pkg.akt.dev/go/node/deployment/v1beta5"
+	dvbeta "pkg.akt.dev/go/node/deployment/v1beta4"
 )
 
 // sum256Seed provides a consistent sha256 value for initial Deployment.Version
@@ -27,12 +27,12 @@ func Deployment(t testing.TB) dtypes.Deployment {
 
 // DeploymentGroup generates a dtype.DepDeploymentGroup in state `GroupOpen`
 // with a set of random required attributes
-func DeploymentGroup(t testing.TB, did dtypes.DeploymentID, gseq uint32) dtypesv1beta4.Group {
+func DeploymentGroup(t testing.TB, did dtypes.DeploymentID, gseq uint32) dvbeta.Group {
 	t.Helper()
-	return dtypesv1beta4.Group{
+	return dvbeta.Group{
 		ID:    dtypes.MakeGroupID(did, gseq),
-		State: dtypesv1beta4.GroupOpen,
-		GroupSpec: dtypesv1beta4.GroupSpec{
+		State: dvbeta.GroupOpen,
+		GroupSpec: dvbeta.GroupSpec{
 			Name:         Name(t, "dgroup"),
 			Requirements: PlacementRequirements(t),
 			Resources:    Resources(t),
@@ -41,9 +41,9 @@ func DeploymentGroup(t testing.TB, did dtypes.DeploymentID, gseq uint32) dtypesv
 }
 
 // GroupSpec generator
-func GroupSpec(t testing.TB) dtypesv1beta4.GroupSpec {
+func GroupSpec(t testing.TB) dvbeta.GroupSpec {
 	t.Helper()
-	return dtypesv1beta4.GroupSpec{
+	return dvbeta.GroupSpec{
 		Name:         Name(t, "dgroup"),
 		Requirements: PlacementRequirements(t),
 		Resources:    Resources(t),
@@ -51,10 +51,10 @@ func GroupSpec(t testing.TB) dtypesv1beta4.GroupSpec {
 }
 
 // DeploymentGroups returns a set of deployment groups generated by DeploymentGroup
-func DeploymentGroups(t testing.TB, did dtypes.DeploymentID, gseq uint32) dtypesv1beta4.Groups {
+func DeploymentGroups(t testing.TB, did dtypes.DeploymentID, gseq uint32) dvbeta.Groups {
 	t.Helper()
 	count := rand.Intn(5) + 5 // nolint:gosec
-	vals := make(dtypesv1beta4.Groups, 0, count)
+	vals := make(dvbeta.Groups, 0, count)
 	for i := 0; i < count; i++ {
 		vals = append(vals, DeploymentGroup(t, did, gseq+uint32(i))) // nolint: gosec
 	}
