+++ go/testutil/v1beta4/deployment.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	dtypes "pkg.akt.dev/go/node/deployment/v1"
-	dvbeta "pkg.akt.dev/go/node/deployment/v1beta4"
+	dtypesv1beta4 "pkg.akt.dev/go/node/deployment/v1beta5"
 )
 
 // sum256Seed provides a consistent sha256 value for initial Deployment.Version
@@ -27,12 +27,12 @@ func Deployment(t testing.TB) dtypes.Deployment {
 
 // DeploymentGroup generates a dtype.DepDeploymentGroup in state `GroupOpen`
 // with a set of random required attributes
-func DeploymentGroup(t testing.TB, did dtypes.DeploymentID, gseq uint32) dvbeta.Group {
+func DeploymentGroup(t testing.TB, did dtypes.DeploymentID, gseq uint32) dtypesv1beta4.Group {
 	t.Helper()
-	return dvbeta.Group{
+	return dtypesv1beta4.Group{
 		ID:    dtypes.MakeGroupID(did, gseq),
-		State: dvbeta.GroupOpen,
-		GroupSpec: dvbeta.GroupSpec{
+		State: dtypesv1beta4.GroupOpen,
+		GroupSpec: dtypesv1beta4.GroupSpec{
 			Name:         Name(t, "dgroup"),
 			Requirements: PlacementRequirements(t),
 			Resources:    Resources(t),
@@ -41,9 +41,9 @@ func DeploymentGroup(t testing.TB, did dtypes.DeploymentID, gseq uint32) dvbeta.
 }
 
 // GroupSpec generator
-func GroupSpec(t testing.TB) dvbeta.GroupSpec {
+func GroupSpec(t testing.TB) dtypesv1beta4.GroupSpec {
 	t.Helper()
-	return dvbeta.GroupSpec{
+	return dtypesv1beta4.GroupSpec{
 		Name:         Name(t, "dgroup"),
 		Requirements: PlacementRequirements(t),
 		Resources:    Resources(t),
@@ -51,10 +51,10 @@ func GroupSpec(t testing.TB) dvbeta.GroupSpec {
 }
 
 // DeploymentGroups returns a set of deployment groups generated by DeploymentGroup
-func DeploymentGroups(t testing.TB, did dtypes.DeploymentID, gseq uint32) dvbeta.Groups {
+func DeploymentGroups(t testing.TB, did dtypes.DeploymentID, gseq uint32) dtypesv1beta4.Groups {
 	t.Helper()
 	count := rand.Intn(5) + 5 // nolint:gosec
-	vals := make(dvbeta.Groups, 0, count)
+	vals := make(dtypesv1beta4.Groups, 0, count)
 	for i := 0; i < count; i++ {
 		vals = append(vals, DeploymentGroup(t, did, gseq+uint32(i))) // nolint: gosec
 	}
