+++ go/testutil/v1beta4/ids.go
@@ -9,7 +9,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	dtypes "pkg.akt.dev/go/node/deployment/v1"
-	mtypes "pkg.akt.dev/go/node/market/v1"
+	mtypes "pkg.akt.dev/go/node/market/v2beta1"
 )
 
 func DeploymentID(t testing.TB) dtypes.DeploymentID {
